requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
pydantic==1.8.2
requests==2.25.1
python-dotenv==0.17.1
pytest==6.2.4
